+++ packages/ng-syz/projects/app/src/app/features/components/docs-example-flow/docs-example-flow-routing.module.ts
@@ -1,0 +1,11 @@
+import { NgModule } from '@angular/core';
+import { RouterModule, Routes } from '@angular/router';
+import { DocsExampleFlowComponent } from './docs-example-flow.component';
+
+const routes: Routes = [{ path: '', component: DocsExampleFlowComponent }];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule],
+})
+export class DocsExampleFlowRoutingModule {}