+++ packages/ng-syz/projects/app/src/app/features/components/docs-example-flow/docs-example-flow.component.spec.ts
@@ -1,0 +1,27 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { NoopAnimationsModule } from '@angular/platform-browser/animations';
+import { SharedModule } from '../../../shared/shared.module';
+
+import { DocsExampleFlowComponent } from './docs-example-flow.component';
+
+describe('DocsExampleFlowComponent', () => {
+  let component: DocsExampleFlowComponent;
+  let fixture: ComponentFixture<DocsExampleFlowComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      imports: [SharedModule, NoopAnimationsModule],
+      declarations: [DocsExampleFlowComponent],
+    }).compileComponents();
+  });
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(DocsExampleFlowComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});