+++ packages/ng-syz/projects/app/src/app/features/components/examples-pages/example-flow-example-two/example-flow-example-two-routing.module.ts
@@ -1,0 +1,13 @@
+import { NgModule } from '@angular/core';
+import { RouterModule, Routes } from '@angular/router';
+import { ExampleFlowExampleTwoComponent } from './example-flow-example-two.component';
+
+const routes: Routes = [
+  { path: '', component: ExampleFlowExampleTwoComponent },
+];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule],
+})
+export class ExampleFlowExampleTwoRoutingModule {}