+++ packages/ng-syz/projects/app/src/app/features/components/examples-pages/examples-pages.component.ts
@@ -1,0 +1,8 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-examples',
+  templateUrl: './examples-pages.component.html',
+  styleUrls: ['./examples-pages.component.scss'],
+})
+export class ExamplesPagesComponent {}