+++ packages/ng-syz/projects/ng-syz/src/public-api.ts
@@ -8,5 +8,5 @@ export * from './lib/cpf-search/cpf-search.component';
 export * from './lib/person-info/person-info.module';
 export * from './lib/person-info/person-info.component';
 
-export * from './lib/login-with-carousel-flow/login-with-carousel-flow.module';
-export * from './lib/login-with-carousel-flow/login-with-carousel-flow.component';
+export * from './lib/example-flow/example-flow.module';
+export * from './lib/example-flow/example-flow.component';