+++ packages/ng-syz/projects/ng-syz/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.4.0](https://github.com/wizsolucoes/syz/compare/@wizsolucoes/ng-syz@0.3.0...@wizsolucoes/ng-syz@0.4.0) (2021-06-16)
+
+
+### Features
+
+* **ng-syz:** add example flow component ([cd9225b](https://github.com/wizsolucoes/syz/commit/cd9225bfa310222a2060a49e72f3109588d1007b))
+
+
+
+
+
 # 0.3.0 (2021-05-28)